Move the TLB code to the persist directory
This is the only plausible use of the TLB.
