+++ opencog/persist/CMakeLists.txt
@@ -7,3 +7,4 @@ ENDIF(HAVE_GEARMAN AND HAVE_GUILE)
 
 ADD_SUBDIRECTORY (sexpr)
 ADD_SUBDIRECTORY (sql)
+ADD_SUBDIRECTORY (tlb)
