+++ opencog/persist/sql/multi-driver/CMakeLists.txt
@@ -20,7 +20,7 @@ ADD_DEPENDENCIES(persist-sql opencog_atom_types atomspace)
 TARGET_LINK_LIBRARIES(persist-sql
 	persist
 	storage-types
-	atomspaceutils
+	tlb
 	atomspace
 )
