+++ opencog/persist/sql/multi-driver/SQLAtomStorage.cc
@@ -34,8 +34,8 @@
 #include <opencog/atoms/atom_types/types.h>
 #include <opencog/atoms/base/Node.h>
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atomspaceutils/TLB.h>
 #include <opencog/persist/storage/storage_types.h>
+#include <opencog/persist/tlb/TLB.h>
 
 #include "SQLAtomStorage.h"
 #include "SQLResponse.h"
