+++ opencog/persist/sql/multi-driver/SQLResponse.h
@@ -30,7 +30,7 @@
 
 #include <opencog/atoms/base/Atom.h>
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atomspaceutils/TLB.h>
+#include <opencog/persist/tlb/TLB.h>
 
 #include "llapi.h"
 #include "SQLAtomStorage.h"
