+++ opencog/persist/sql/multi-driver/SQLSpaces.cc
@@ -24,7 +24,7 @@
 #include <unistd.h>
 
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atomspaceutils/TLB.h>
+#include <opencog/persist/tlb/TLB.h>
 
 #include "SQLAtomStorage.h"
 #include "SQLResponse.h"
