+++ tests/persist/CMakeLists.txt
@@ -1,5 +1,6 @@
 ADD_SUBDIRECTORY (sexpr)
 ADD_SUBDIRECTORY (sql)
+ADD_SUBDIRECTORY (tlb)
 
 IF (HAVE_GUILE AND HAVE_GEARMAN)
    ADD_SUBDIRECTORY (gearman)
