+++ tests/persist/tlb/CMakeLists.txt
@@ -1,0 +1,7 @@
+
+LINK_LIBRARIES(
+	atombase
+	atomspace
+)
+
+ADD_CXXTEST(TLBUTest)
