+++ tests/persist/tlb/TLBUTest.cxxtest
@@ -1,5 +1,5 @@
 /*
- * tests/atomspace/TLBUTest.cxxtest
+ * tests/persist/tlb/TLBUTest.cxxtest
  *
  * Copyright (C) 2011 OpenCog Foundation
  * All Rights Reserved
@@ -26,7 +26,7 @@
 #include <stdio.h>
 
 #include <opencog/atoms/base/Node.h>
-#include <opencog/atomspaceutils/TLB.h>
+#include <opencog/persist/tlb/TLB.h>
 #include <opencog/atoms/atom_types/atom_types.h>
 #include <opencog/util/exceptions.h>
